requirements.txt
requests~=2.32.4
python-dotenv~=1.1.1
bs4~=0.0.2
beautifulsoup4~=4.13.4
python-telegram-bot~=22.2